refactor(plugin): extract full/{kernel_lifecycle,kernel_construction,render_target,methods}.rs (#1092)
Split gpu_context/mod.rs's FullAccess vtable section into per-banner
submodules under full/, mirroring the limited/ layout.

* full/mod.rs — re-exports + the drop_handle defensive no-op stub
  (kept at the FullAccess root because GpuContextFullAccess::Drop
  dispatches on the struct's handle_kind discriminator directly
  without routing through this vtable slot).

* full/kernel_lifecycle.rs (~340 LoC, Linux-only) — Arc-handle
  clone/drop pairs for every kernel β-shape the cdylib carries:
  compute / graphics / ray-tracing kernel, texture ring, color
  converter (v4 #917), acceleration structure (v4 #917), command
  recorder (v5 #984).

* full/kernel_construction.rs (~370 LoC, Linux-only) —
  create_compute_kernel, create_graphics_kernel,
  create_ray_tracing_kernel, create_texture_ring. Each validates
  scope_token + decodes the cdylib's msgpack descriptor + runs the
  privileged engine constructor.

* full/render_target.rs (~120 LoC) — Phase C3
  acquire_render_target_dma_buf_image (DRM-modifier-tiled DMA-BUF
  VkImage allocation through the host's privileged RHI path).

* full/methods.rs (~1080 LoC) — Phase D (#906) privileged methods:
  wait_device_idle, acquire_output_texture,
  upload_pixel_buffer_as_texture, color_converter,
  create_command_recorder, build_triangles_blas, build_tlas,
  supports_ray_tracing_pipeline, gpu_capabilities,
  create_timeline_semaphore, import_dma_buf_storage_buffer,
  check_in_surface, host_vulkan_device_arc,
  host_vulkan_texture_arc.

Parent gpu_context/mod.rs now carries only the GpuContext-handle
clone/drop callbacks + the two vtable statics + accessors + test
modules. Test modules will be relocated in a follow-up commit.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
